requirements.txt
absl-py==0.9.0
Brotli==1.0.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.5.0
future==0.18.2
google-auth==1.19.2
google-auth-oauthlib==0.4.1
grpcio==1.30.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
llvmlite==0.33.0
Markdown==3.2.2
MarkupSafe==1.1.1
numba==0.50.1
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Quaternion==3.5.2.post4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
threadpoolctl==2.1.0
torch==1.5.1
torchvision==0.6.1
ujson==3.0.0
umap==0.1.1
umap-learn==0.4.6
urllib3==1.25.10
Werkzeug==1.0.1
